Added Prime95 support and changed Multi-thread recognition
Finally found it is possible to launch the Prime95 torture test from command line, so have added support for that to the Python script.
Changed how the threading of the test (single or multi-threaded) is identified. Rather than by test name in Output.r, it should be set with the test's function in the CPU Thermal.py (variable MULTI), and it is then passed to Input.r for Output.r to read. It is better to have this controlled when the test is configured in the Python script than after the fact in the R scripts.