Move to Environment instead of CSV
Instead of using a CSV for storing all of the data, using saveRDS and an environment allows the formatting of a data frame to be saved and loaded as well. The CSV will still be made, but using the environment will likely be better in the long run.
It is easy to go back though.